requirements.txt
fiftyone==1.5.2
fiftyone_brain==0.21.2
fiftyone_db==1.1.7
imgui==2.0.0
matplotlib==3.10.3
numpy==2.3.0
pandas==2.3.0
Pillow==11.2.1
pygame==2.6.1
pygame_gui==0.6.14
pyglm==2.8.2
pymeshlab==2023.12.post3
PyOpenGL==3.1.9
PyOpenGL==3.1.9
skimage==0.0
ultralytics==8.3.150
